#330 FilteringParserDelegate - missing end marker handling fixed
