+++ src/main/java/com/fasterxml/jackson/core/filter/FilteringParserDelegate.java
@@ -233,13 +233,9 @@ public class FilteringParserDelegate extends JsonParserDelegate
         // Check for _allowMultipleMatches - false and at least there is one token - which is _currToken
         // check for no buffered context _exposedContext - null
         // If all the conditions matches then check for scalar / non-scalar property
-        
+
         if (!_allowMultipleMatches && (_currToken != null) && (_exposedContext == null)) {
-            // if not scalar and ended successfully, then return null
-            if ((_currToken.isStructEnd() && _headContext.isStartHandled()) ){
-                return (_currToken = null);
-            }
-            // else if scalar, and scalar not present in obj/array and !includePath and INCLUDE_ALL
+            // if scalar, and scalar not present in obj/array and !includePath and INCLUDE_ALL
             // matched once, return null
             if (_currToken.isScalarValue() && !_headContext.isStartHandled() && !_includePath
                     && (_itemFilter == TokenFilter.INCLUDE_ALL)) {
